requirements.txt
aiohttp==3.7.4.post0
discord==1.7.3
tenacity==8.0.1
gql[all]==3.0.0a6
python-dotenv==0.19.0
pydantic==1.8.2
